refactor: prevents mixed server/clients imports
