+++ packages/app-store/routing-forms/lib/getSerializableForm.ts
@@ -1,7 +1,7 @@
 import type { App_RoutingForms_Form } from "@prisma/client";
 import type { z } from "zod";
 
-import { entityPrismaWhereClause } from "@calcom/lib/entityPermissionUtils";
+import { entityPrismaWhereClause } from "@calcom/lib/entityPermissionUtils.server";
 import logger from "@calcom/lib/logger";
 import { safeStringify } from "@calcom/lib/safeStringify";
 import { RoutingFormSettings } from "@calcom/prisma/zod-utils";
