+++ packages/app-store/routing-forms/trpc/getResponseWithFormFields.handler.ts
@@ -1,6 +1,6 @@
 import type { z } from "zod";
 
-import { canAccessEntity } from "@calcom/lib/entityPermissionUtils";
+import { canAccessEntity } from "@calcom/lib/entityPermissionUtils.server";
 import { getTranslation } from "@calcom/lib/server/i18n";
 import { prisma } from "@calcom/prisma";
 import type { TrpcSessionUser } from "@calcom/trpc/server/types";
